Move `State` to `stateful` :tada:
