+++ packages/engine/src/datastore/table/mod.rs
@@ -2,6 +2,6 @@
 //! This is about a single simulation's data, conceptually we have Tables for agent state, context,
 //! message state
 pub mod context;
-pub mod state;
+pub mod create_remove;
 pub mod sync;
 pub mod task_shared_store;
